# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentació de la nostra API",
        "description": "",
        "contact": {
            "name": "Equip de Spottunes",
            "email": "a20pedgarguz@inspedralbes.cat"
        },
        "license": {
            "name": "Licencia API",
            "url": "http://www.example.com/licenses/api"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/events/all": {
            "get": {
                "tags": [
                    "Esdeveniments"
                ],
                "summary": "Obtenir llista completa d'esdeveniments",
                "description": "Obté una llista completa d'esdeveniments on el camp 'artist' no és nul, ordenats per data i hora.",
                "operationId": "getAllEvents",
                "responses": {
                    "200": {
                        "description": "Llista d'esdeveniments obtinguda amb èxit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "events": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "artist": {
                                                        "type": "string",
                                                        "example": "Artista Exemple"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-06-01"
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "format": "time",
                                                        "example": "19:00:00"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Lloc Exemple"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No s'han trobat esdeveniments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No s'han trobat esdeveniments"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}